% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localrank-package.r
\docType{package}
\name{localrank-package}
\alias{localrank-package}
\title{localrank}
\description{
localrank
}
\references{
Programming with Big Data in R Website: \url{https://pbdr.org/}
}
\keyword{Package}
